refactor(frontend): centralize click outside logic
